Create some seeds for local development
This adds some database seeds so you can start
the server locally and actually get some data out
and see the response.

I added Faker to generate some of the data.

There is also an environment variable that
can be set to control the number of messages that
are seeded.